+++ crates/language-server/src/logging.rs
