+++ crates/language-server/src/main.rs
@@ -1,9 +1,8 @@
 mod backend;
-mod cli_args;
+mod cli;
 mod functionality;
-mod logger;
+mod logging;
 mod lsp_actor;
-mod lsp_actor_service;
 mod lsp_streams;
 mod server;
 mod util;
@@ -17,7 +16,7 @@ use async_lsp::panic::CatchUnwindLayer;
 use async_lsp::server::LifecycleLayer;
 use async_lsp::tracing::TracingLayer;
 use clap::Parser;
-use cli_args::{CliArgs, Commands};
+use cli::{CliArgs, Commands};
 use futures::io::AsyncReadExt;
 use futures::StreamExt;
 use futures_net::TcpListener;
