+++ crates/language-server/src/server.rs
@@ -1,9 +1,9 @@
 use std::ops::ControlFlow;
 
 use crate::functionality::handlers::{FileChange, FilesNeedDiagnostics, NeedsDiagnostics};
-use crate::logger;
-use crate::lsp_actor::{LspActor, LspDispatcher};
-use crate::lsp_actor_service::LspActorService;
+use crate::logging;
+use crate::lsp_actor::service::LspActorService;
+use crate::lsp_actor::LspActor;
 use crate::lsp_streams::RouterStreams;
 use act_locally::builder::ActorBuilder;
 use async_lsp::lsp_types::notification::{
@@ -20,7 +20,6 @@ use tracing::{error, info};
 
 use crate::backend::Backend;
 use crate::functionality::{goto, handlers};
-use act_locally::actor::HandlerRegistration;
 use async_lsp::router::Router;
 use async_lsp::{
     lsp_types::request::Initialize,
@@ -33,36 +32,35 @@ pub(crate) struct Lol;
 
 pub(crate) fn setup(client: ClientSocket, name: String) -> BoxLspService<Value, ResponseError> {
     info!("Setting up server");
-    let worker_thread_client = client.clone();
-    let logger_client = client.clone();
-    let actor_ref = ActorBuilder::new()
+    let client_for_actor = client.clone();
+    let client_for_logging = client.clone();
+    let backend_actor = ActorBuilder::new()
         .with_name(name)
-        .with_state_init(move || Ok(Backend::new(worker_thread_client)))
-        .with_subscriber_init(logger::setup(logger_client))
+        .with_state_init(move || Ok(Backend::new(client_for_actor)))
+        .with_subscriber_init(logging::setup(client_for_logging))
         .spawn();
-    let mut dispatcher = LspDispatcher::new();
+    let got_actor_ref = backend_actor.is_ok();
 
-    let got_actor_ref = actor_ref.is_ok();
-    let mut services = match actor_ref {
-        Ok(actor_ref) => {
-            HandlerRegistration {
-                actor_ref: &actor_ref,
-                dispatcher: &mut dispatcher,
-            }
-            // .handle_event::<Lol>(handle_lol)
-            .handle_request::<Initialize>(handlers::initialize)
-            .handle_notification::<Initialized>(handlers::initialized)
-            .handle_request::<GotoDefinition>(goto::handle_goto_definition)
-            .handle_request::<HoverRequest>(handlers::handle_hover_request)
-            .handle_notification::<DidOpenTextDocument>(handlers::handle_did_open_text_document)
-            .handle_notification::<DidChangeTextDocument>(handlers::handle_did_change_text_document)
-            .handle_notification::<DidChangeWatchedFiles>(handlers::handle_did_change_watched_files)
-            .handle_event::<FileChange>(handlers::handle_file_change)
-            .handle_event::<FilesNeedDiagnostics>(handlers::handle_files_need_diagnostics)
-            .handle_notification::<notification::Exit>(handlers::handle_exit);
+    let mut services = match backend_actor {
+        Ok(backend_actor_ref) => {
+            let mut lsp_actor_service = LspActorService::new(backend_actor_ref.clone());
+            lsp_actor_service
+                .handle_request::<Initialize>(handlers::initialize)
+                .handle_notification::<Initialized>(handlers::initialized)
+                .handle_request::<GotoDefinition>(goto::handle_goto_definition)
+                .handle_request::<HoverRequest>(handlers::handle_hover_request)
+                .handle_notification::<DidOpenTextDocument>(handlers::handle_did_open_text_document)
+                .handle_notification::<DidChangeTextDocument>(
+                    handlers::handle_did_change_text_document,
+                )
+                .handle_notification::<DidChangeWatchedFiles>(
+                    handlers::handle_did_change_watched_files,
+                )
+                .handle_event::<FileChange>(handlers::handle_file_change)
+                .handle_event::<FilesNeedDiagnostics>(handlers::handle_files_need_diagnostics)
+                .handle_notification::<notification::Exit>(handlers::handle_exit);
 
-            let actor_service = LspActorService::new(actor_ref.clone(), dispatcher);
-            vec![BoxLspService::new(actor_service)]
+            vec![BoxLspService::new(lsp_actor_service)]
         }
         Err(_) => Vec::new(),
     };
